pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "manytask-monorepo"
version = "0.1.0"
description = "Manytask system with integrated Checker"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    {include = "manytask"},
    {include = "checker", from = "checker"}
]

[tool.poetry.scripts]
checker = "checker.__main__:cli"

[tool.poetry.dependencies]
python = "^3.12"
# Manytask dependencies
gunicorn = "23.0.0"
flask = "3.1.1"
cachelib = "0.13.0"
werkzeug = "3.1.3"
flask-sqlalchemy = "3.1.1"
psycopg2-binary = "2.9.10"
gspread = ">=5.0.0,!=6.0.0,<7.0.0"
python-gitlab = "6.4.0"
authlib = "1.6.5"
python-dotenv = "1.1.0"
pyyaml = "6.0.1"
requests = "2.32.4"
pydantic = "2.11.1"
pytz = {version = ">=2022.0,<2023.4", python = "<3.9"}
alembic = "1.16.1"
flask-wtf = "^1.2.2"
# Checker dependencies
Jinja2 = ">=3.0.0,<4.0.0"
click = ">=8.0,<9.0"
rich = ">=13.0.0,<15.0.0"
urllib3 = ">=2.0.0,<3.0.0"
GitPython = ">=3.0.0,<4.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "0.13.1"
mypy = "1.18.2"
isort = "6.0.1"
black = "25.1.0"
types-requests = "2.32.0.20240907"
types-PyYAML = "6.0.12.12"
pytest = "8.4.0"
pytest-cov = "7.0.0"
pytest-mock = ">=3.0.0,<4.0.0"
pytest-xdist = "^3.8.0"
requests-mock = ">=1.0.0,<2.0.0"
pre-commit = "4.3.0"
testcontainers = "4.13.0"
bs4 = "^0.0.2"

[tool.mypy]
no_incremental = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
allow_untyped_decorators = true
exclude = [
    "tests/",
    "checker/tests/",
    ".tmp/",
    ".venv/",
    "manytask/migrations/versions"
]


[tool.pyright]
include = ["manytask", "checker"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    ".venv",
    "venv",
    ".tmp",
    "tests",
    "checker/tests"
]


[tool.pytest.ini_options]
minversion = "6.0"
python_files = "test_*.py"
addopts = "--cov=manytask/ --cov=checker/ --cov-report=term-missing --cov-report=xml"
python_paths = ["."]
markers = [
    "firejail: tests that require firejail (Linux-only sandbox tool)",
    "skip_integration: skip integration tests",
    "skip_unit: skip unit tests",
    "skip_doctest: skip doctests",
]


[tool.ruff]
line-length = 120
target-version = "py312"
src = ["manytask", "checker", "tests", "checker/tests"]

# Exclude patterns
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    "*.ipynb",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".tmp",
    "manytask/migrations/versions"
]

[tool.ruff.lint]
# Combine both configurations' rules
select = ["E", "F", "I", "TID252", "TID253", "C901", "PLR"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.mccabe]
# Maximum cyclomatic complexity allowed
max-complexity = 15

[tool.ruff.lint.pylint]
max-nested-blocks = 5
max-args = 10

[tool.ruff.lint.isort]
known-first-party = ["manytask", "checker", "tests"]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120
lines_after_imports = 2
combine_as_imports = true
default_section = "THIRDPARTY"
known_first_party = ["manytask", "checker", "tests"]
